+++ chains/solana/utils/mcms/mcm.go
@@ -1,9 +1,11 @@
 package mcms
 
 import (
+	crypto_rand "crypto/rand"
 	"encoding/binary"
 	"errors"
 	"fmt"
+	"time"
 
 	"github.com/gagliardetto/solana-go"
 
@@ -78,6 +80,30 @@ func GetSeenSignedHashesPDA(msigName [32]byte, root [32]byte, validUntil uint32)
 		validUntilBytes,
 	}, config.McmProgram)
 	return pda
+}
+
+// utils for padding mcm id
+func PadString32(input string) ([32]byte, error) {
+	var result [32]byte
+	inputBytes := []byte(input)
+	inputLen := len(inputBytes)
+	if inputLen > 32 {
+		return result, errors.New("input string exceeds 32 bytes")
+	}
+	startPos := 32 - inputLen
+	copy(result[startPos:], inputBytes)
+	return result, nil
+}
+
+func UnpadString32(input [32]byte) string {
+	startPos := 0
+	for i := 0; i < len(input); i++ {
+		if input[i] != 0 {
+			startPos = i
+			break
+		}
+	}
+	return string(input[startPos:])
 }
 
 type McmConfigArgs struct {
@@ -157,13 +183,10 @@ func GetNewMcmMultisig(name [32]byte) Multisig {
 func GetPreloadSignersIxs(signerAddresses [][20]uint8, msigName [32]byte, multisigCfgPDA solana.PublicKey, cfgSignersPDA solana.PublicKey, authority solana.PublicKey, appendChunkSize int) ([]solana.Instruction, error) {
 	ixs := make([]solana.Instruction, 0)
 
-	parsedTotalSigners, pErr := SafeToUint8(len(signerAddresses))
-	if pErr != nil {
-		return nil, pErr
-	}
 	initSignersIx, isErr := mcm.NewInitSignersInstruction(
 		msigName,
-		parsedTotalSigners,
+		//nolint:gosec
+		uint8(len(signerAddresses)),
 		multisigCfgPDA,
 		cfgSignersPDA,
 		authority,
@@ -224,16 +247,12 @@ func GetAppendSignersIxs(signerAddresses [][20]uint8, msigName [32]byte, multisi
 func GetMcmPreloadSignaturesIxs(signatures []mcm.Signature, msigName [32]byte, root [32]uint8, validUntil uint32, signaturesPDA solana.PublicKey, authority solana.PublicKey, appendChunkSize int) ([]solana.Instruction, error) {
 	ixs := make([]solana.Instruction, 0)
 
-	parsedTotalSigs, pErr := SafeToUint8(len(signatures))
-	if pErr != nil {
-		return nil, pErr
-	}
-
 	initSigsIx, isErr := mcm.NewInitSignaturesInstruction(
 		msigName,
 		root,
 		validUntil,
-		parsedTotalSigs,
+		//nolint:gosec
+		uint8(len(signatures)),
 		signaturesPDA,
 		authority,
 		solana.SystemProgramID,
@@ -403,4 +422,22 @@ func IxToMcmTestOpNode(multisig solana.PublicKey, msigSigner solana.PublicKey, i
 	}
 
 	return node, nil
+}
+
+// simple salt generator that uses the current Unix timestamp(in mills)
+func SimpleSalt() ([32]byte, error) {
+	var salt [32]byte
+	now := time.Now().UnixMilli()
+	if now < 0 {
+		return salt, fmt.Errorf("negative timestamp: %d", now)
+	}
+	// unix timestamp in millseconds
+	binary.BigEndian.PutUint64(salt[:8], uint64(now))
+	// Next 8 bytes: Crypto random
+	randBytes := make([]byte, 8)
+	if _, err := crypto_rand.Read(randBytes); err != nil {
+		return salt, fmt.Errorf("failed to generate random bytes: %w", err)
+	}
+	copy(salt[8:16], randBytes)
+	return salt, nil
 }
